model_service/requirements.txt
# Data fetching dependencies
requests>=2.28.0
python-dotenv>=1.0.0
httpx>=0.24.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# ML/Deep Learning
torch>=2.0.0
pytorch-forecasting>=1.0.0
lightning>=2.0.0
